Breaking: Remove all references to the dust bundler
